+++ sync/pkg/server.go
@@ -3,19 +3,13 @@ package sync
 import (
 	"context"
 	"crypto/tls"
-	"encoding/json"
 	"fmt"
 	"log"
 	"net"
-	"os"
-	"time"
+	"sync"
 
-	"github.com/fsnotify/fsnotify"
-	"golang.org/x/exp/maps"
-
-	"buf.build/gen/go/open-feature/flagd/grpc/go/sync/v1/syncv1grpc"
-	v1 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/sync/v1"
-	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
+	"buf.build/gen/go/open-feature/flagd/grpc/go/flagd/sync/v1/syncv1grpc"
+	v1 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/flagd/sync/v1"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 )
@@ -25,12 +19,6 @@ type Server struct {
 }
 
 func (s *Server) Start() {
-	err := SetupTraceProvider()
-	if err != nil {
-		log.Printf("Error setting up telemetry : %s\n", err.Error())
-		return
-	}
-
 	listen, err := net.Listen("tcp", s.Config.Host+":"+s.Config.Port)
 	if err != nil {
 		log.Printf("Error when listening to address : %s\n", err.Error())
@@ -43,10 +31,13 @@ func (s *Server) Start() {
 		return
 	}
 
-	options = append(options, grpc.StreamInterceptor(otelgrpc.StreamServerInterceptor()))
-
 	server := grpc.NewServer(options...)
 	sync, err := NewSyncImpl(s.Config.Files.Array)
+	if err != nil {
+		log.Printf("Error configuring the server : %s\n", err.Error())
+		return
+	}
+
 	syncv1grpc.RegisterFlagSyncServiceServer(server, &sync)
 
 	fmt.Printf("Server listening : %s\n", s.Config.Host+":"+s.Config.Port)
@@ -75,101 +66,71 @@ func (s *Server) buildOptions() ([]grpc.ServerOption, error) {
 
 // SyncImpl implements the flagd Sync contract
 type SyncImpl struct {
-	filePaths []string
-	watcher   *fsnotify.Watcher
+	fw *fileWatcher
 }
 
 func NewSyncImpl(filePaths []string) (SyncImpl, error) {
-	watcher, err := fsnotify.NewWatcher()
-	for _, filePath := range filePaths {
-
-		if err != nil {
-			return SyncImpl{}, err
-		}
-		watcher.Add(filePath)
+	fw := &fileWatcher{
+		paths: filePaths,
+		subs:  make(map[interface{}]chan<- []byte),
+		mu:    sync.Mutex{},
 	}
+
+	err := fw.init()
+	if err != nil {
+		log.Printf("Error starting file watcher %s\n", err.Error())
+		return SyncImpl{}, err
+	}
+
 	return SyncImpl{
-		filePaths,
-		watcher,
+		fw,
 	}, nil
 }
 
 func (s *SyncImpl) SyncFlags(req *v1.SyncFlagsRequest, stream syncv1grpc.FlagSyncService_SyncFlagsServer) error {
-	log.Printf("Requesting flags for provider : %s", req.ProviderId)
+	log.Printf("Requesting flags for provider: %s\n", req.ProviderId)
 
-	marshalled, err := s.readFlags()
-	if err != nil {
-		log.Println("error reading flags:", err)
-	}
-	err = stream.Send(&v1.SyncFlagsResponse{
-		FlagConfiguration: string(marshalled),
-		State:             v1.SyncState_SYNC_STATE_ALL,
+	// initial read
+	err := stream.Send(&v1.SyncFlagsResponse{
+		FlagConfiguration: string(s.fw.getCurrentData()),
 	})
 	if err != nil {
-		log.Println("error sending initial stream:", err)
+		log.Printf("Error sending initial stream: %v\n", err)
+		return err
 	}
+
+	listener := make(chan []byte)
+	s.fw.subscribe(listener)
 
 	// Start listening for events.
 	for {
 		select {
-		case event, ok := <-s.watcher.Events:
-			if !ok {
-				log.Println("unable to process file event")
+		case data := <-listener:
+			err = stream.Send(&v1.SyncFlagsResponse{
+				FlagConfiguration: string(data),
+			})
+			if err != nil {
+				// this is probably a close
+				message := fmt.Sprintf("error sending stream, likely closed: %v", err)
+				log.Println(message)
+				return nil
 			}
-			if event.Has(fsnotify.Write) {
-				marshalled, err := s.readFlags()
-				if err != nil {
-					log.Println("error reading flags:", err)
-				}
-				err = stream.Send(&v1.SyncFlagsResponse{
-					FlagConfiguration: string(marshalled),
-					State:             v1.SyncState_SYNC_STATE_ALL,
-				})
-				if err != nil {
-					log.Println("error sending stream:", err)
-				}
-			}
-		case err, _ := <-s.watcher.Errors:
-			log.Println("error in file watcher:", err)
+		case <-stream.Context().Done():
+			s.fw.unSubscribe(listener)
+			log.Printf("Stream completed for provider: %s\n", req.ProviderId)
+			return nil
 		}
 	}
-}
-
-func (s *SyncImpl) readFlags() ([]byte, error) {
-	flags := make(map[string]any)
-
-	for _, path := range s.filePaths {
-		bytes, err := os.ReadFile(path)
-		for len(bytes) == 0 {
-			// this is a fitly hack
-			// file writes are NOT atomic and often when they are occur they have transitional empty states
-			// this "re-reads" the file in these cases a bit later
-			time.Sleep(10 * time.Millisecond)
-			bytes, err = os.ReadFile(path)
-		}
-		if err != nil {
-			return nil, err
-		}
-		parsed := make(map[string]map[string]any)
-		json.Unmarshal(bytes, &parsed)
-		maps.Copy(flags, parsed["flags"])
-	}
-
-	payload := make(map[string]any)
-	payload["flags"] = flags
-	marshalled, err := json.Marshal(payload)
-	if err != nil {
-		return nil, err
-	}
-	return marshalled, nil
 }
 
 func (s *SyncImpl) FetchAllFlags(context.Context, *v1.FetchAllFlagsRequest) (*v1.FetchAllFlagsResponse, error) {
-	marshalled, err := s.readFlags()
-	if err != nil {
-		log.Println("error reading flags:", err)
-	}
+	marshalled := s.fw.getCurrentData()
+
 	return &v1.FetchAllFlagsResponse{
 		FlagConfiguration: string(marshalled),
 	}, nil
+}
+
+func (s *SyncImpl) GetMetadata(context.Context, *v1.GetMetadataRequest) (*v1.GetMetadataResponse, error) {
+	return &v1.GetMetadataResponse{}, nil
 }